Renaming component to GitHub API CLI
